# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * CodeQuest API specification
 * OpenAPI spec version: 0.2.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
  message?: string;
}

export interface RegisterRequest {
  /**
   * @minLength 3
   * @maxLength 30
   */
  username: string;
  email: string;
  /** @minLength 6 */
  password: string;
}

export interface LoginRequest {
  email: string;
  password: string;
}

export type BadgeCategory = (typeof BadgeCategory)[keyof typeof BadgeCategory];

export const BadgeCategory = {
  solving: "solving",
  xp: "xp",
  streak: "streak",
  quest: "quest",
  level: "level",
} as const;

export type BadgeRarity = (typeof BadgeRarity)[keyof typeof BadgeRarity];

export const BadgeRarity = {
  common: "common",
  rare: "rare",
  epic: "epic",
  legendary: "legendary",
} as const;

export interface Badge {
  slug: string;
  name: string;
  description: string;
  icon: string;
  category: BadgeCategory;
  rarity: BadgeRarity;
  earned: boolean;
  earnedAt?: string | null;
}

export interface AllBadgesResponse {
  badges: Badge[];
}

export interface MyBadgesResponse {
  badges: Badge[];
  totalEarned: number;
}

export type DailyQuestResponseDifficulty =
  (typeof DailyQuestResponseDifficulty)[keyof typeof DailyQuestResponseDifficulty];

export const DailyQuestResponseDifficulty = {
  Easy: "Easy",
  Medium: "Medium",
  Hard: "Hard",
} as const;

export interface DailyQuestResponse {
  problemId: number;
  problemTitle: string;
  difficulty: DailyQuestResponseDifficulty;
  topic: string;
  bonusXp: number;
  date: string;
  completed: boolean;
}

export interface UserProfile {
  id: number;
  username: string;
  email: string;
  xp: number;
  level: number;
  starRank: number;
  solvedCount: number;
  streak: number;
  dailyQuestsCompleted: number;
  rank?: number | null;
  badges: Badge[];
  createdAt: string;
}

export interface AuthResponse {
  token: string;
  user: UserProfile;
}

export type ProblemDifficulty =
  (typeof ProblemDifficulty)[keyof typeof ProblemDifficulty];

export const ProblemDifficulty = {
  Easy: "Easy",
  Medium: "Medium",
  Hard: "Hard",
} as const;

export interface Problem {
  id: number;
  title: string;
  slug: string;
  difficulty: ProblemDifficulty;
  topic: string;
  tags: string[];
  acceptanceRate?: number | null;
  solvedCount: number;
  isDailyQuest: boolean;
}

export interface ProblemExample {
  input: string;
  output: string;
  explanation?: string | null;
}

export interface StarterCode {
  python: string;
  javascript: string;
}

export type ProblemDetail = Problem & {
  description: string;
  constraints?: string | null;
  examples: ProblemExample[];
  starterCode: StarterCode;
  bonusXp: number;
};

export interface ProblemsListResponse {
  problems: Problem[];
  total: number;
  page: number;
  limit: number;
}

export type SubmitCodeRequestLanguage =
  (typeof SubmitCodeRequestLanguage)[keyof typeof SubmitCodeRequestLanguage];

export const SubmitCodeRequestLanguage = {
  python: "python",
  javascript: "javascript",
} as const;

export interface SubmitCodeRequest {
  problemId: number;
  code: string;
  language: SubmitCodeRequestLanguage;
}

export type RunCodeRequestLanguage =
  (typeof RunCodeRequestLanguage)[keyof typeof RunCodeRequestLanguage];

export const RunCodeRequestLanguage = {
  python: "python",
  javascript: "javascript",
} as const;

export interface RunCodeRequest {
  problemId: number;
  code: string;
  language: RunCodeRequestLanguage;
}

export interface TestCaseResult {
  passed: boolean;
  input: string;
  expectedOutput: string;
  actualOutput?: string | null;
  error?: string | null;
  time?: number | null;
}

export type SubmissionResultStatus =
  (typeof SubmissionResultStatus)[keyof typeof SubmissionResultStatus];

export const SubmissionResultStatus = {
  Accepted: "Accepted",
  Wrong_Answer: "Wrong Answer",
  Time_Limit_Exceeded: "Time Limit Exceeded",
  Runtime_Error: "Runtime Error",
  Compilation_Error: "Compilation Error",
} as const;

export interface SubmissionResult {
  id: number;
  status: SubmissionResultStatus;
  passedCount: number;
  totalCount: number;
  xpEarned: number;
  bonusXpEarned: number;
  results: TestCaseResult[];
  runtime?: number | null;
  newBadges: Badge[];
  streakUpdated: boolean;
  newStreak: number;
  newStarRank?: number | null;
  isDailyQuest: boolean;
  levelUp: boolean;
  newLevel?: number | null;
}

export interface RunResult {
  results: TestCaseResult[];
  allPassed: boolean;
}

export interface Submission {
  id: number;
  problemId: number;
  problemTitle: string;
  language: string;
  status: string;
  passedCount: number;
  totalCount: number;
  runtime?: number | null;
  submittedAt: string;
}

export interface SubmissionHistoryResponse {
  submissions: Submission[];
}

export interface LeaderboardEntry {
  rank: number;
  userId: number;
  username: string;
  xp: number;
  level: number;
  starRank: number;
  solvedCount: number;
  streak: number;
}

export interface LeaderboardResponse {
  entries: LeaderboardEntry[];
  total: number;
  page: number;
  limit: number;
  currentUserRank?: number | null;
}

export interface WeeklyLeaderboardEntry {
  rank: number;
  userId: number;
  username: string;
  solvedThisWeek: number;
  xpThisWeek: number;
  level: number;
  starRank: number;
}

export interface WeeklyLeaderboardResponse {
  entries: WeeklyLeaderboardEntry[];
  total: number;
  page: number;
  limit: number;
  weekStart: string;
  weekEnd: string;
  currentUserRank?: number | null;
}

export type GetProblemsParams = {
  difficulty?: GetProblemsDifficulty;
  topic?: string;
  search?: string;
  page?: number;
  limit?: number;
};

export type GetProblemsDifficulty =
  (typeof GetProblemsDifficulty)[keyof typeof GetProblemsDifficulty];

export const GetProblemsDifficulty = {
  Easy: "Easy",
  Medium: "Medium",
  Hard: "Hard",
} as const;

export type GetSubmissionHistoryParams = {
  problemId?: number;
};

export type GetLeaderboardParams = {
  page?: number;
  limit?: number;
};

export type GetWeeklyLeaderboardParams = {
  page?: number;
  limit?: number;
};