# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.2.0
  description: CodeQuest API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/register:
    post:
      operationId: register
      tags: [auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterRequest"
      responses:
        "201":
          description: User created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Conflict
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Current user profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /problems:
    get:
      operationId: getProblems
      tags: [problems]
      parameters:
        - name: difficulty
          in: query
          schema:
            type: string
            enum: [Easy, Medium, Hard]
        - name: topic
          in: query
          schema:
            type: string
        - name: search
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: List of problems
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProblemsListResponse"
  /problems/{id}:
    get:
      operationId: getProblem
      tags: [problems]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Problem detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /leaderboard:
    get:
      operationId: getLeaderboard
      tags: [leaderboard]
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
      responses:
        "200":
          description: Global leaderboard
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LeaderboardResponse"
  /leaderboard/weekly:
    get:
      operationId: getWeeklyLeaderboard
      tags: [leaderboard]
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
      responses:
        "200":
          description: Weekly leaderboard (problems solved this week)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WeeklyLeaderboardResponse"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
      required: [error]
    RegisterRequest:
      type: object
      properties:
        username:
          type: string
          minLength: 3
          maxLength: 30
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 6
      required: [username, email, password]
    LoginRequest:
      type: object
      properties:
        email:
          type: string
          format: email
        password:
          type: string
      required: [email, password]
    Badge:
      type: object
      properties:
        slug:
          type: string
        name:
          type: string
        description:
          type: string
        icon:
          type: string
        category:
          type: string
          enum: [solving, xp, streak, quest, level]
        rarity:
          type: string
          enum: [common, rare, epic, legendary]
        earned:
          type: boolean
        earnedAt:
          type: string
          format: date-time
          nullable: true
      required: [slug, name, description, icon, category, rarity, earned]
...
    UserProfile:
      type: object
      properties:
        id:
          type: integer
        username:
          type: string
        email:
          type: string
        xp:
          type: integer
        level:
          type: integer
        starRank:
          type: integer
        solvedCount:
          type: integer
        streak:
          type: integer
        dailyQuestsCompleted:
          type: integer
        rank:
          type: integer
          nullable: true
        badges:
          type: array
          items:
            $ref: "#/components/schemas/Badge"
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - username
        - email
        - xp
        - level
        - starRank
        - solvedCount
        - streak
        - dailyQuestsCompleted
        - badges
        - createdAt
    AuthResponse:
      type: object
      properties:
        token:
          type: string
        user:
          $ref: "#/components/schemas/UserProfile"
      required: [token, user]
    Problem:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        slug:
          type: string
        difficulty:
          type: string
          enum: [Easy, Medium, Hard]
        topic:
          type: string
        tags:
          type: array
          items:
            type: string
        acceptanceRate:
          type: number
          nullable: true
        solvedCount:
          type: integer
        isDailyQuest:
          type: boolean
      required: [id, title, slug, difficulty, topic, tags, solvedCount, isDailyQuest]
    ProblemDetail:
      allOf:
        - $ref: "#/components/schemas/Problem"
        - type: object
          properties:
            description:
              type: string
            constraints:
              type: string
              nullable: true
            examples:
              type: array
              items:
                $ref: "#/components/schemas/ProblemExample"
            starterCode:
              $ref: "#/components/schemas/StarterCode"
            bonusXp:
              type: integer
          required: [description, examples, starterCode, bonusXp]
    ProblemExample:
      type: object
      properties:
        input:
          type: string
        output:
          type: string
        explanation:
          type: string
          nullable: true
      required: [input, output]
    StarterCode:
      type: object
      properties:
        python:
          type: string
        javascript:
          type: string
      required: [python, javascript]
    ProblemsListResponse:
      type: object
      properties:
        problems:
          type: array
          items:
            $ref: "#/components/schemas/Problem"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
      required: [problems, total, page, limit]
...
    LeaderboardEntry:
      type: object
      properties:
        rank:
          type: integer
        userId:
          type: integer
        username:
          type: string
        xp:
          type: integer
        level:
          type: integer
        starRank:
          type: integer
        solvedCount:
          type: integer
        streak:
          type: integer
      required: [rank, userId, username, xp, level, starRank, solvedCount, streak]
    LeaderboardResponse:
      type: object
      properties:
        entries:
          type: array
          items:
            $ref: "#/components/schemas/LeaderboardEntry"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
        currentUserRank:
          type: integer
          nullable: true
      required: [entries, total, page, limit]
    WeeklyLeaderboardEntry:
      type: object
      properties:
        rank:
          type: integer
        userId:
          type: integer
        username:
          type: string
        solvedThisWeek:
          type: integer
        xpThisWeek:
          type: integer
        level:
          type: integer
        starRank:
          type: integer
      required: [rank, userId, username, solvedThisWeek, xpThisWeek, level, starRank]
    WeeklyLeaderboardResponse:
      type: object
      properties:
        entries:
          type: array
          items:
            $ref: "#/components/schemas/WeeklyLeaderboardEntry"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
        weekStart:
          type: string
        weekEnd:
          type: string
        currentUserRank:
          type: integer
          nullable: true
      required: [entries, total, page, limit, weekStart, weekEnd]