# lib/api-zod/src/generated/types/leaderboardResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * CodeQuest API specification
 * OpenAPI spec version: 0.2.0
 */
import type { LeaderboardEntry } from "./leaderboardEntry";

export interface LeaderboardResponse {
  entries: LeaderboardEntry[];
  total: number;
  page: number;
  limit: number;
  currentUserRank?: number | null;
}