# lib/api-zod/src/generated/types/problem.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * CodeQuest API specification
 * OpenAPI spec version: 0.2.0
 */
import type { ProblemDifficulty } from "./problemDifficulty";

export interface Problem {
  id: number;
  title: string;
  slug: string;
  difficulty: ProblemDifficulty;
  topic: string;
  tags: string[];
  acceptanceRate?: number | null;
  solvedCount: number;
  isDailyQuest: boolean;
}