# lib/api-zod/src/generated/types/problemDetail.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * CodeQuest API specification
 * OpenAPI spec version: 0.2.0
 */
import type { Problem } from "./problem";
import type { ProblemExample } from "./problemExample";
import type { StarterCode } from "./starterCode";

export type ProblemDetail = Problem & {
  description: string;
  constraints?: string | null;
  examples: ProblemExample[];
  starterCode: StarterCode;
  bonusXp: number;
};