# lib/api-zod/src/generated/types/problemExample.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * CodeQuest API specification
 * OpenAPI spec version: 0.2.0
 */

export interface ProblemExample {
  input: string;
  output: string;
  explanation?: string | null;
}