# lib/api-zod/src/generated/types/problemsListResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * CodeQuest API specification
 * OpenAPI spec version: 0.2.0
 */
import type { Problem } from "./problem";

export interface ProblemsListResponse {
  problems: Problem[];
  total: number;
  page: number;
  limit: number;
}