# lib/api-zod/src/generated/types/weeklyLeaderboardResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * CodeQuest API specification
 * OpenAPI spec version: 0.2.0
 */
import type { WeeklyLeaderboardEntry } from "./weeklyLeaderboardEntry";

export interface WeeklyLeaderboardResponse {
  entries: WeeklyLeaderboardEntry[];
  total: number;
  page: number;
  limit: number;
  weekStart: string;
  weekEnd: string;
  currentUserRank?: number | null;
}